Refactoring Part 1. Moved CommonFuncs -> StringFuncs and added all file/directory related functions to FSFuncs which has been moved under include directory from include/ProjectManagement
