+++ include/FSFuncs.hpp
@@ -3,12 +3,16 @@
 
 #include <string>
 
+#include "PackageManagement/PackageData.hpp"
+
 void SetFolderPaths( std::string & directory,
 		     const std::string & projectname,
 		     std::string & projfolder,
 		     std::string & srcfolder,
 		     std::string & includefolder,
 		     std::string & buildfolder );
+
+bool DirExists( const std::string & dir );
 
 int CreateDir( const std::string & dir );
 
@@ -20,5 +24,11 @@ bool IsLatestBuild( std::string filename );
 long long GetLastModifiedTime( std::string file );
 
 int GetFilesInDir( std::string dir, std::vector< std::string > & temp, bool recursive = false );
+
+bool CheckNecessaryPermissions( const Package & pkg, bool framework_exists = false );
+
+bool CreateArchiveDir( const Package & pkg );
+
+std::string GetArchiveDir( const Package & pkg );
 
 #endif // FSFUNCS_HPP
