+++ include/ProjectManagement/ProjectFileGenerator.hpp
@@ -2,7 +2,6 @@
 #define PROJECTFILEGENERATOR_HPP
 
 #include "ProjectData.hpp"
-#include "FSFuncs.hpp"
 #include "ConfigMgr.hpp"
 
 int GenerateProjectFiles( ProjectData & data );
