+++ include/StringFuncs.hpp
