+++ src/ProjectManagement/ProjectBuilder.cpp
@@ -2,7 +2,8 @@
 #include <vector>
 #include <string>
 
-#include "../../include/CommonFuncs.hpp"
+#include "../../include/StringFuncs.hpp"
+
 #include "../../include/ProjectManagement/ProjectData.hpp"
 #include "../../include/ProjectManagement/ConfigMgr.hpp"
 #include "../../include/ProjectManagement/BuildFilesGenerator.hpp"
